+++ CHANGELOG.md
@@ -1,3 +1,7 @@
 ## [0.1.0] — UNRELEASED
 
-- Initial release
+### Added
+
+#### Gems
+
+- `Gem::Author` for authors info to be used primarily in gem specs.